changing torch.tensor to torch.full_like for MIN_MASKED_ATTENTION_VALUE
(cherry picked from commit 806f2cfda824fd160f6cb3d0bf1d433b17924a65)

+++ QEfficient/transformers/models/codegen/modeling_codegen.py
@@ -49,7 +49,7 @@ class QEffCodeGenAttention(CodeGenAttention):
 
         # Need to be a tensor, otherwise we get error: `RuntimeError: expected scalar type float but found double`.
         # Need to be on the same device, otherwise `RuntimeError: ..., x and y to be on the same device`
-        mask_value = torch.full_like(attn_weights, MIN_MASKED_ATTENTION_VALUE)
+        mask_value = torch.full_like(attn_weights, MIN_MASKED_ATTENTION_VALUE, dtype=attn_weights.dtype)
 
         if attention_mask is not None:
             # Apply the attention mask
